refactor scripts
