--- scripts/3-deploy-derived-price-feeds.ts
@@ -1,7 +1,0 @@
-// Should be able to specify array of symbols for which you need to run this script
-
-// Deploy price feeds using appropritae salts from config. Could be a way to fetch salt using the symbol.
-
-// If price feed is supposed to be derived then go ahead and deploy the derived feed
-
-// TODO: passing the array of symbols and chainId - based on chainId config file is fetched, only run on required symbols
