--- scripts/4-set-token-oracles.ts
@@ -1,11 +1,0 @@
-// Connect and return instance with given EOA signer and provided address
-
-// Should be able to specify array of symbols for which you need to run this script
-
-// Fetch price feed addresses and priceFeedFunction
-
-// Set token oracles using the current owner
-
-// Transfer ownership back to intended owner if ownership was transferred back to run the script
-
-// TODO: passing the array of symbols and chainId - based on chainId config file is fetched, only run on required symbols
