+++ test/foundry/token-paymaster/TokenPaymasterMumbai.t.sol
@@ -165,7 +165,6 @@ contract TokenPaymasterMumbaiTest is SATestBase {
 
         entryPoint.handleOps(arraifyOps(op), dan.addr);
 
-        // todo // review fails to validate updated balances
         console2.log("paymaster balance after ", usdc.balanceOf(address(_btpm)));
         assertNotEq(usdc.balanceOf(address(sa)), 100e6);
     }
